refactor: Extract runSecurityScan, generateSecurityReport, extractResourceNames to security.go
main.go: 1587 → 1341 lines
